Initial check-in of some unit tests for ECQuiz.
The tests can be run like this:

    $INSTANCE_HOME/bin/zopectl test --libdir $INSTANCE_HOME/Products/ECQuiz

This command changes starting with Zope 2.9, which requires you to do
this:

    $INSTANCE_HOME/bin/zopectl test -s Products.ECQuiz

Supposedly, the tests are run in a newly created Plone site and each
transaction that occurs in the test code is automatically aborted, so
your ZODB should be unchanged after the tests have run.  Having a
separate Zope site dedicated to unit-testing might still be a good
idea, though.

Running the tests should produce lots of debug messages and
deprecation warnings and finally something like:

    Adding ECQuiz ... done (0.414s)
    .....
    ----------------------------------------------------------------------
    Ran 5 tests in 2.186s

    OK

For me, this is followed by a traceback, which I think is actually the
expected behavior:

    Traceback (most recent call last):
      File "/opt/zope-2.8.7-final/lib/python/Zope2/Startup/zopectl.py",
      line 322, in ?
        main()
      File "/opt/zope-2.8.7-final/lib/python/Zope2/Startup/zopectl.py",
      line 283, in main
        c.onecmd(" ".join(options.args))
      File "/opt/python-2.3.5/lib/python2.3/cmd.py", line 210, in onecmd
        return func(arg)
      File "/opt/zope-2.8.7-final/lib/python/Zope2/Startup/zopectl.py",
      line 268, in do_test
        os.waitpid(pid, 0)
    OSError: [Errno 4] Interrupted system call

For more info, see <http://plone.org/documentation/tutorial/testing/>,
especially sections 3, »Running tests« and 4, »Writing unit tests.«